refactor: Moved standalone components to the shared directory
