+++ src/BurstChat.Client/src/app/components/root/root.component.html
@@ -1,3 +1,2 @@
 <router-outlet></router-outlet>
-
 <burst-popup-list></burst-popup-list>
