+++ src/BurstChat.Client/src/app/components/root/root.component.ts
@@ -1,8 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { NotifyService } from 'src/app/services/notify/notify.service';
-import { PopupListComponent } from 'src/app/components/popup-list/popup-list.component';
+import { PopupListComponent } from 'src/app/components/shared/popup-list/popup-list.component';
 
 /**
  * This class represents an angular component that is the primary host component for all other displayed components.
@@ -16,7 +15,6 @@ import { PopupListComponent } from 'src/app/components/popup-list/popup-list.com
     standalone: true,
     imports: [
         RouterOutlet,
-        FontAwesomeModule,
         PopupListComponent
     ]
 })
