+++ src/BurstChat.Client/src/app/components/shared/avatar/avatar.component.html
