+++ src/BurstChat.Client/src/app/components/shared/avatar/avatar.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/theme.scss';
+@import '../../../../scss/theme';
 
 .avatar {
     width: 35px;
