+++ src/BurstChat.Client/src/app/components/shared/avatar/avatar.component.spec.ts
