+++ src/BurstChat.Client/src/app/components/shared/card-body/card-body.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/theme.scss';
+@import '../../../../scss/theme';
 
 .burst-chat-body {
     min-height: 60px;
