+++ src/BurstChat.Client/src/app/components/shared/card-body/card-body.component.ts
