+++ src/BurstChat.Client/src/app/components/shared/card-footer/card-footer.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/theme.scss';
+@import '../../../../scss/theme';
 
 .burst-chat-footer {
     background-color: transparent;
