+++ src/BurstChat.Client/src/app/components/shared/card-footer/card-footer.component.spec.ts
