+++ src/BurstChat.Client/src/app/components/shared/card-header/card-header.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/theme.scss';
+@import '../../../../scss/theme';
 
 .burst-chat-header {
     background-color: $dark-color;
