+++ src/BurstChat.Client/src/app/components/shared/card/card.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/theme.scss';
+@import '../../../../scss/theme';
 
 .burst-card {
     background-color: $secondary-color;
