+++ src/BurstChat.Client/src/app/components/shared/card/card.component.ts
