+++ src/BurstChat.Client/src/app/components/shared/dialog/dialog.component.scss
@@ -1,5 +1,5 @@
-@import '../../../scss/theme';
-@import '../../../scss/colors';
+@import '../../../../scss/theme';
+@import '../../../../scss/colors';
 
 .dialog-container {
     @extend .blurred-color;
