+++ src/BurstChat.Client/src/app/components/shared/dialog/dialog.component.ts
