+++ src/BurstChat.Client/src/app/components/shared/expander/expander.component.scss
@@ -1,6 +1,6 @@
-@import '../../../scss/theme';
-@import '../../../scss/sizing';
-@import '../../../scss/text';
+@import '../../../../scss/theme';
+@import '../../../../scss/sizing';
+@import '../../../../scss/text';
 
 .expander {
     @extend .w-100;
