+++ src/BurstChat.Client/src/app/components/shared/expander/expander.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { faAngleRight, faAngleDown } from '@fortawesome/free-solid-svg-icons';
-import { ShimmerListComponent } from 'src/app/components/shimmer-list/shimmer-list.component';
+import { ShimmerListComponent } from 'src/app/components/shared/shimmer-list/shimmer-list.component';
 
 /**
  * This class represents an angular component that displays an expandable area.
