+++ src/BurstChat.Client/src/app/components/shared/image-crop/image-crop.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/theme';
+@import '../../../../scss/theme';
 
 input[type="file"] {
     display: none;
