+++ src/BurstChat.Client/src/app/components/shared/image-crop/image-crop.component.ts
