+++ src/BurstChat.Client/src/app/components/shared/popup-list/popup-list.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/theme';
+@import '../../../../scss/theme';
 
 .popup-list-container {
     position: fixed;
