+++ src/BurstChat.Client/src/app/components/shared/popup/popup.component.scss
