+++ src/BurstChat.Client/src/app/components/shared/popup/popup.component.ts
@@ -1,6 +1,6 @@
 import { NgClass } from '@angular/common';
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
-import { FaIconLibrary, FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 import { PopupMessage } from 'src/app/models/notify/popup-message';
