+++ src/BurstChat.Client/src/app/components/shared/shimmer-list/shimmer-list.component.html
