+++ src/BurstChat.Client/src/app/components/shared/shimmer-list/shimmer-list.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/theme';
+@import '../../../../scss/theme';
 
 @-webkit-keyframes placeholderShimmer {
     0% {
