+++ src/BurstChat.Client/src/app/modules/burst/burst.module.ts
@@ -21,13 +21,13 @@ import { UserComponent } from 'src/app/modules/burst/components/user/user.compon
 import { IncomingCallComponent } from 'src/app/modules/burst/components/incoming-call/incoming-call.component';
 import { OngoingCallComponent } from 'src/app/modules/burst/components/ongoing-call/ongoing-call.component';
 import { AddServerComponent } from 'src/app/modules/burst/components/add-server/add-server.component';
-import { DialogComponent } from 'src/app/components/dialog/dialog.component';
-import { CardComponent } from 'src/app/components/card/card.component';
-import { CardHeaderComponent } from 'src/app/components/card-header/card-header.component';
-import { CardBodyComponent } from 'src/app/components/card-body/card-body.component';
-import { CardFooterComponent } from 'src/app/components/card-footer/card-footer.component';
-import { AvatarComponent } from 'src/app/components/avatar/avatar.component';
-import { ExpanderComponent } from 'src/app/components/expander/expander.component';
+import { DialogComponent } from 'src/app/components/shared/dialog/dialog.component';
+import { CardComponent } from 'src/app/components/shared/card/card.component';
+import { CardHeaderComponent } from 'src/app/components/shared/card-header/card-header.component';
+import { CardBodyComponent } from 'src/app/components/shared/card-body/card-body.component';
+import { CardFooterComponent } from 'src/app/components/shared/card-footer/card-footer.component';
+import { AvatarComponent } from 'src/app/components/shared/avatar/avatar.component';
+import { ExpanderComponent } from 'src/app/components/shared/expander/expander.component';
 
 import { library } from '@fortawesome/fontawesome-svg-core';
 import {
