+++ src/BurstChat.Client/src/app/modules/user/user.module.ts
@@ -10,14 +10,14 @@ import { EditUserComponent } from 'src/app/modules/user/components/edit-user/edi
 import { EditUserInfoComponent } from 'src/app/modules/user/components/edit-user-info/edit-user-info.component';
 import { EditUserInvitationsComponent } from 'src/app/modules/user/components/edit-user-invitations/edit-user-invitations.component';
 import { EditUserMediaComponent } from 'src/app/modules/user/components/edit-user-media/edit-user-media.component';
-import { CardComponent } from 'src/app/components/card/card.component';
-import { CardHeaderComponent } from 'src/app/components/card-header/card-header.component';
-import { CardBodyComponent } from 'src/app/components/card-body/card-body.component';
-import { CardFooterComponent } from 'src/app/components/card-footer/card-footer.component';
-import { ImageCropComponent } from 'src/app/components/image-crop/image-crop.component';
-import { DialogComponent } from 'src/app/components/dialog/dialog.component';
-import { AvatarComponent } from 'src/app/components/avatar/avatar.component';
-import { ExpanderComponent } from 'src/app/components/expander/expander.component';
+import { CardComponent } from 'src/app/components/shared/card/card.component';
+import { CardHeaderComponent } from 'src/app/components/shared/card-header/card-header.component';
+import { CardBodyComponent } from 'src/app/components/shared/card-body/card-body.component';
+import { CardFooterComponent } from 'src/app/components/shared/card-footer/card-footer.component';
+import { ImageCropComponent } from 'src/app/components/shared/image-crop/image-crop.component';
+import { DialogComponent } from 'src/app/components/shared/dialog/dialog.component';
+import { AvatarComponent } from 'src/app/components/shared/avatar/avatar.component';
+import { ExpanderComponent } from 'src/app/components/shared/expander/expander.component';
 
 @NgModule({
     imports: [
